Refactor firebase and image picker functions into separate files
- Create a folder for helper classes
- Create a class to hold all firebase manipulation functions
- Create a class to hold image picker functions
- Add missing comments to the new files
- Refactor pages to use the helper classes
